+++ devicemanagement/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- Updated `setModuleFrequency` test to revert back to original frequencies and
+  check by setting a list of test frequencies.
+- Renamed `setModuleFrequency` to `setAndGetModuleFrequency` because now this
+  test sets the frequency and validates it by reading back those frequencies.
 ### Deprecated
 ### Removed
 ### Fixed